refactor: split main.py into routers/ — predict, simulation, trade, market
Pure structural refactor — zero behaviour changes. Extracted all route
handlers and their private helpers out of the 1792-line main.py into four
focused APIRouter modules:

  backend/routers/predict.py    — /health, /sparklines, /debug, /predict, /compare
  backend/routers/simulation.py — /simulation/suggest, /simulation/performance, /simulation/state (CRUD)
  backend/routers/trade.py      — /trade-setup, /chat
  backend/routers/market.py     — /dcf/{symbol}

backend/dependencies.py holds all service singletons (InfluxService,
ForecastStore, SerpService, YFinanceService, AnalystJuryService,
SentimentService) so routers share one instance each without importing
from main.py (avoids circular imports).

main.py is now a thin 50-line entry point: lifespan, exception handler,
router registration, uvicorn __main__ block.

ruff: all checks passed. pytest: 22/22 passed.
