+++ backend/routers/market.py
@@ -1,0 +1,105 @@
+# backend/routers/market.py
+import asyncio
+import logging
+
+from fastapi import APIRouter, HTTPException
+
+from dependencies import yf_svc
+
+router = APIRouter()
+logger = logging.getLogger(__name__)
+
+# ---------------------------------------------------------------------------
+# DCF Intrinsic Value
+# ---------------------------------------------------------------------------
+RISK_FREE_RATE      = 0.045   # ~10Y Treasury yield
+EQUITY_RISK_PREMIUM = 0.055   # Damodaran
+
+
+def _run_dcf(
+    fcf: float,
+    growth: float,
+    wacc: float,
+    terminal_growth: float = 0.025,
+    years: int = 5,
+) -> float:
+    """Returns equity value (total PV of FCF stream + terminal value)."""
+    pv = 0.0
+    cf = fcf
+    for yr in range(1, years + 1):
+        cf *= (1 + growth)
+        pv += cf / ((1 + wacc) ** yr)
+    # Gordon Growth terminal value
+    tv = cf * (1 + terminal_growth) / max(wacc - terminal_growth, 0.001)
+    pv += tv / ((1 + wacc) ** years)
+    return pv
+
+
+@router.get("/dcf/{symbol}")
+async def dcf_valuation(symbol: str):
+    import yfinance as yf
+
+    info = await asyncio.to_thread(yf_svc.fetch_info, symbol)
+
+    # Raw data
+    fcf           = info.get("free_cash_flow")
+    beta          = info.get("beta")
+    revenue_growth = info.get("revenue_growth")
+    current_price = info.get("current_price", 0)
+
+    # Shares outstanding — fetch separately
+    ticker_obj = await asyncio.to_thread(lambda: yf.Ticker(symbol.upper()))
+    ticker_info = await asyncio.to_thread(lambda: ticker_obj.info)
+    shares = (
+        ticker_info.get("sharesOutstanding")
+        or ticker_info.get("impliedSharesOutstanding")
+        or 1
+    )
+
+    # Validate: need positive FCF and shares
+    if not fcf or not isinstance(fcf, (int, float)) or fcf <= 0 or not shares or shares <= 0:
+        raise HTTPException(
+            status_code=422,
+            detail="Insufficient data for DCF (negative/zero FCF or missing shares)",
+        )
+
+    # WACC
+    b = float(beta) if beta and beta != "N/A" else 1.0
+    b = max(0.1, min(b, 3.0))
+    wacc_base = RISK_FREE_RATE + b * EQUITY_RISK_PREMIUM
+
+    # Growth rate
+    g = float(revenue_growth) if revenue_growth and revenue_growth != "N/A" else 0.05
+    g = max(-0.20, min(g, 0.40))   # clamp to ±20-40%
+
+    TERMINAL_GROWTH = 0.025
+
+    def scenario(wacc_delta: float, growth_mult: float) -> dict:
+        w  = max(wacc_base + wacc_delta, TERMINAL_GROWTH + 0.001)
+        gr = g * growth_mult
+        total_pv  = _run_dcf(float(fcf), gr, w, TERMINAL_GROWTH)
+        per_share = round(total_pv / shares, 2)
+        upside    = (
+            round((per_share - float(current_price)) / float(current_price) * 100, 1)
+            if current_price else 0
+        )
+        return {
+            "wacc":            round(w  * 100, 2),
+            "growth_rate":     round(gr * 100, 2),
+            "intrinsic_value": per_share,
+            "upside_pct":      upside,
+        }
+
+    logger.info(f"[DCF] {symbol.upper()} — FCF={fcf:.0f} beta={b:.2f} wacc_base={wacc_base:.3f} g={g:.3f}")
+    return {
+        "symbol":             symbol.upper(),
+        "current_price":      round(float(current_price), 2),
+        "bear":               scenario(+0.02, 0.7),
+        "base":               scenario(0.0,   1.0),
+        "bull":               scenario(-0.02, 1.3),
+        "shares_outstanding": int(shares),
+        "fcf_billions":       round(float(fcf) / 1e9, 2),
+        "wacc_base":          round(wacc_base * 100, 2),
+        "growth_rate_base":   round(g * 100, 2),
+        "method":             "FCF",
+    }
